Update configuring-ibmi-vm.md
We have received several cases from customers where the control key does not work work on the keyboard.  An RFE is open to development for this:  https://www.ibm.com/developerworks/rfe/execute?use_case=viewRfe&CR_ID=139386.    I have found that sometimes it works and sometimes it does not.  When it does not, the reset button on the UI performs the same function.   I updated the fork to reflect that.